Finish TODO;s, including readme, rename manifest to manifests
