+++ cli/manifests.ts
